requirements.txt
lxml
python-dotenv
requests
beautifulsoup4
pandas
numpy
scikit-learn
kneed